[ALPS06544280] scsi: ufs: update skhpb
update skhpb to 1.5.8

MTK-Commit-Id: bd5e4b4b7840680c2df007ccb0172b2d8383b0f0

CR-Id: ALPS06544280
Feature: ufs

+++ drivers/scsi/ufs/ufshpb_skh.h
@@ -352,7 +352,7 @@ struct skhpb_lu {
 	/* for debug constant variables */
 	unsigned long long lu_num_blocks;
 
-	u8 lun;
+	int lun;
 
 	struct ufs_hba *hba;
